refactor: update package imports to use the new test and infrastructure structure
