+++ internal/handlers/auth_handler_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/fifawcp/api/internal/dtos"
 	"github.com/fifawcp/api/internal/infrastructure/config"
 	"github.com/fifawcp/api/internal/infrastructure/validator"
-	"github.com/fifawcp/api/internal/packages/mocks"
-	"github.com/fifawcp/api/internal/packages/testutils"
+	"github.com/fifawcp/api/internal/test/mocks"
+	"github.com/fifawcp/api/internal/test/testutils"
 	"github.com/go-chi/chi/v5"
 	"github.com/stretchr/testify/assert"
 )
