+++ internal/handlers/debug_handler.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/fifawcp/api/internal/infrastructure/config"
-	"github.com/fifawcp/api/internal/packages/httputils"
-	"github.com/fifawcp/api/internal/packages/totp"
+	"github.com/fifawcp/api/internal/httputils"
+	"github.com/fifawcp/api/internal/infrastructure/totp"
 	"github.com/go-chi/chi/v5"
 )
