+++ internal/handlers/errors.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/fifawcp/api/internal/domain"
 	"github.com/fifawcp/api/internal/infrastructure/logging"
-	httputils "github.com/fifawcp/api/internal/packages/httputils"
+	httputils "github.com/fifawcp/api/internal/httputils"
 	"github.com/go-chi/chi/v5/middleware"
 )
