+++ internal/handlers/oauth_handler.go
@@ -6,7 +6,7 @@ import (
 	"github.com/fifawcp/api/internal/infrastructure/config"
 	"github.com/fifawcp/api/internal/infrastructure/logging"
 	"github.com/fifawcp/api/internal/httpctx"
-	"github.com/fifawcp/api/internal/packages/httputils"
+	"github.com/fifawcp/api/internal/httputils"
 	"github.com/fifawcp/api/internal/services"
 )
