+++ internal/handlers/user_handler.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/fifawcp/api/internal/infrastructure/logging"
 	"github.com/fifawcp/api/internal/httpctx"
-	"github.com/fifawcp/api/internal/packages/httputils"
+	"github.com/fifawcp/api/internal/httputils"
 	"github.com/fifawcp/api/internal/services"
 )
