+++ internal/middlewares/rate_limit_middleware.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/fifawcp/api/internal/infrastructure/logging"
 	"github.com/fifawcp/api/internal/infrastructure/ratelimit"
-	"github.com/fifawcp/api/internal/packages/httputils"
+	"github.com/fifawcp/api/internal/httputils"
 )
 
 func RateLimitByIP(
