+++ internal/middlewares/require_board_membership_middleware.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/fifawcp/api/internal/domain"
 	"github.com/fifawcp/api/internal/httpctx"
-	"github.com/fifawcp/api/internal/packages/httputils"
+	"github.com/fifawcp/api/internal/httputils"
 	"github.com/fifawcp/api/internal/services"
 	"github.com/go-chi/chi/v5"
 	"github.com/google/uuid"
