+++ internal/middlewares/require_valid_match_id_middleware.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/fifawcp/api/internal/httpctx"
 	"github.com/fifawcp/api/internal/infrastructure/logging"
-	"github.com/fifawcp/api/internal/packages/httputils"
+	"github.com/fifawcp/api/internal/httputils"
 	"github.com/go-chi/chi/v5"
 )
