+++ internal/middlewares/validate_oauth_callback_middleware.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/fifawcp/api/internal/httpctx"
 	"github.com/fifawcp/api/internal/infrastructure/logging"
-	"github.com/fifawcp/api/internal/packages/httputils"
+	"github.com/fifawcp/api/internal/httputils"
 )
 
 func ValidateOAuthCallback(logger logging.Logger) func(next http.Handler) http.Handler {
