+++ internal/services/auth_service_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/fifawcp/api/internal/dtos"
 	"github.com/fifawcp/api/internal/infrastructure/auth"
 	"github.com/fifawcp/api/internal/infrastructure/config"
-	"github.com/fifawcp/api/internal/packages/mocks"
-	"github.com/fifawcp/api/internal/packages/totp"
+	"github.com/fifawcp/api/internal/test/mocks"
+	"github.com/fifawcp/api/internal/infrastructure/totp"
 	"github.com/stretchr/testify/assert"
 )
